fixed issues across platforms
-------------ssues acros This line and the following will be ignored --------------

modified:
  include/mariadb++/connection.hpp
  include/mariadb++/data.hpp
  include/mariadb++/date_time.hpp
  include/mariadb++/decimal.hpp
  include/mariadb++/exceptions.hpp
  include/mariadb++/last_error.hpp
  include/mariadb++/mariadb++.hpp
  include/mariadb++/result_set.hpp
  include/mariadb++/save_point.hpp
  include/mariadb++/statement.hpp
  include/mariadb++/time.hpp
  include/mariadb++/time_span.hpp
  include/mariadb++/transaction.hpp
  include/mariadb++/types.hpp
  src/account.cpp
  src/bind.cpp
  src/bind.hpp
  src/concurrency.cpp
  src/config.h.in
  src/connection.cpp
  src/date_time.cpp
  src/decimal.cpp
  src/exceptions.cpp
  src/last_error.cpp
  src/private.cpp
  src/private.hpp
  src/result_set.cpp
  src/save_point.cpp
  src/statement.cpp
  src/time.cpp
  src/transaction.cpp
  src/worker.cpp
  src/worker.hpp
  unittest/CMakeLists.txt
  unittest/database.cpp
  unittest/date_time.cpp
  unittest/decimal.cpp
  unittest/main.cpp
  unittest/test.cpp
  unittest/test.hpp
unknown:
  .DS_Store
  include/mariadb++/query.hpp
  src/.DS_Store
  src/query.cpp
  unittest/.DS_Store

+++ include/mariadb++/data.hpp
@@ -48,7 +48,7 @@ namespace mariadb
 			m_size(0),
 			m_data(0)
 		{
-			create(data, count);
+			create(data, static_cast<u32>(count));
 		}
 
 		//
